Break up the compare module
